+++ streampipes-extensions/streampipes-connectors-opcua/src/main/java/org/apache/streampipes/extensions/connectors/opcua/adapter/utils/ExceptionMessageExtractor.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.streampipes.connect.iiot.adapters.opcua.utils;
+package org.apache.streampipes.extensions.connectors.opcua.adapter.utils;
 
 import org.eclipse.milo.opcua.stack.core.UaException;
