+++ streampipes-extensions/streampipes-connectors-opcua/src/main/java/org/apache/streampipes/extensions/connectors/opcua/adapter/utils/OpcUaNodeVariants.java
@@ -16,7 +16,8 @@
  *
  */
 
-package org.apache.streampipes.connect.iiot.adapters.opcua.utils;
+package org.apache.streampipes.extensions.connectors.opcua.adapter.utils;
+
 
 import jakarta.annotation.Nullable;
