+++ streampipes-extensions/streampipes-connectors-opcua/src/main/java/org/apache/streampipes/extensions/connectors/opcua/sink/OpcUaParameters.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.streampipes.sinks.databases.jvm.opcua;
+package org.apache.streampipes.extensions.connectors.opcua.sink;
 
 public final class OpcUaParameters {
   private final String hostname;
