+++ streampipes-extensions/streampipes-sinks-databases-jvm/pom.xml
@@ -31,6 +31,11 @@
         <!-- StreamPipes dependencies -->
         <dependency>
             <groupId>org.apache.streampipes</groupId>
+            <artifactId>streampipes-connectors-opcua</artifactId>
+            <version>0.93.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.streampipes</groupId>
             <artifactId>streampipes-sdk-bundle</artifactId>
             <version>0.93.0-SNAPSHOT</version>
         </dependency>
